backend/pyproject.toml
[project]
name = "stockpulse-backend"
version = "0.1.0"
description = "StockPulse — AI-Powered Stock Tracking & Insights Platform (Backend)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "tenacity>=8.2.3",
    "anthropic>=0.26.0",
    "redis>=5.0.3",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "types-passlib>=1.7.7",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends.legacy:build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
strict = true
ignore_missing_imports = true
python_version = "3.12"
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
